Last commit for now.
It’s still not working perfectly, but I want to move on. I realized
that to make it work like a real calculator (which is more than what is
required or even mentioned as extra credit in the homework assignment)
would be a bigger task than I want to dedicate time to right now. A
real calculator would, for example, make 4+4/2 be 6 instead of 4. I’d
rather move on with the course and even leave the things I could
accomplish easily, like making the .0s disappear in the description.