Description and display work properly.
Added NumberFormatter extension.

Multiplication and division first doesn’t work though. Instead, the
results are always shown immediately: Entering “4 + 2 x" already shows
6 in the display.
I tried the built-in iPhone calculator: When you tap times or divided
and there is a + or - calculation pending, the result is not yet shown
in the display. In all other cases, the result is shown immediately. So
this behavior should be my goal, if I decide I want to make this
calculator be good.